Move CRD fixup hack for 3.11
Due to an OLM bug in OCP 3.11, the generated CRD needs some tweaking.
Previously this was being done via a bespoke script invoked through a
go:generate directive. After boilerplate adoption, this script was doing
`operator-sdk generate` redundantly (it was already part of the
`op-generate` target).

This commit removes the script and the redundant `operator-sdk
generate`s, and adds a temporary `make` target to do the CRD fixup. All
of this will go away once we no longer need to run on 3.11.